part1 done

+++ src/ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: yejeong <yejeong@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/27 12:14:18 by yejeong           #+#    #+#             */
-/*   Updated: 2021/05/10 13:19:31 by yejeong          ###   ########.fr       */
+/*   Updated: 2021/05/10 17:10:46 by yejeong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-unsigned int	ft_strlcpy(char *dest, char *src, size_t size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	if (size != 0)
